renamed files for pytest autodiscovery and organized testing utils
