+++ tests/evolution/selection_strategy_test.py
@@ -3,7 +3,7 @@ import pytest
 from watts.evolution.selection_strategy import SelectRandomly
 from watts.pair.agent_environment_pair import Pairing
 
-from tests.test_classes import MockGenerator, MockPair, MockSolver
+from tests.utils.test_classes import MockGenerator, MockPair, MockSolver
